+++ src/epics/layers.js
@@ -6,8 +6,6 @@ import { addLayer, updateLayer, loadLayer } from '../actions/layers';
 import { errorActionCreator } from '../actions/helpers';
 import { fetchLayer } from '../loaders/layers';
 import { drillUpDown } from '../util/map';
-import { getPeriodFromFilters } from '../util/analytics';
-import { getRelativePeriods } from '../util/periods';
 
 const isNewLayer = config => config.id === undefined;
 
@@ -40,25 +38,4 @@ export const drillLayer = (action$, store) =>
         )
         .mergeMap(config => [closeContextMenu(), loadLayer(config)]);
 
-// Reload layers with relative periods using a specific date (from interpretation)
-export const setRelativePeriodDate = (action$, store) =>
-    action$
-        .ofType(types.MAP_RELATIVE_PERIOD_DATE_SET)
-        .mergeMap(action =>
-            store
-                .getState()
-                .map.mapViews.filter(config => {
-                    const period = getPeriodFromFilters(config.filters);
-                    return (
-                        period &&
-                        getRelativePeriods().find(p => p.id === period.id)
-                    );
-                })
-                .map(config => ({
-                    ...config,
-                    relativePeriodDate: action.payload,
-                }))
-        )
-        .map(loadLayer);
-
-export default combineEpics(loadLayerEpic, drillLayer, setRelativePeriodDate);
+export default combineEpics(loadLayerEpic, drillLayer);